# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Transferências e Usuários",
    "version": "1.0.0",
    "description": "API para registro, login, consulta de usuários e transferências."
  },
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": { "type": "string" },
                  "password": { "type": "string" },
                  "name": { "type": "string" },
                  "favorecidos": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["login", "password", "name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário registrado" },
          "400": { "description": "Erro de validação" }
        }
      }
    },
    "/users/login": {
      "post": {
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "login": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": ["login", "password"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login realizado" },
          "400": { "description": "Campos obrigatórios ausentes" },
          "401": { "description": "Login ou senha inválidos" }
        }
      }
    },
    "/users": {
      "get": {
        "summary": "Listar usuários",
        "responses": {
          "200": { "description": "Lista de usuários" }
        }
      }
    },
    "/transfers": {
      "post": {
        "summary": "Realizar transferência",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "from": { "type": "string" },
                  "to": { "type": "string" },
                  "amount": { "type": "number" }
                },
                "required": ["from", "to", "amount"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Transferência realizada" },
          "400": { "description": "Erro de validação" }
        }
      },
      "get": {
        "summary": "Listar transferências",
        "responses": {
          "200": { "description": "Lista de transferências" }
        }
      }
    }
  }
}